per email 7/6/2022:   As I wrote before, the tool the NEFSC has been using to meet the HQ security requirement on repositories became a paid tool last month.  NEFSC is in talks with the developers to try to arrive at a price point that meets our needs.  In the meantime, all NEFSC repos that are being actively used must re-enable (if you turned it off) and update your yaml code as follows.  I have updated the NEFSC-Template with the new code, so this only applies to repos already in our GitHub.com organization.
   In the file named "secretScan.yml" located in .github/workflows, change the last line from:
uses: zricethezav/gitleaks-action@master to
uses: gitleaks/gitleaks-action@v1.6.0

This will use the last free version of gitleaks.  Note that this may change in the future, if the NEFSC is able to buy a commercial product or find a better open source alternative.  But in the meantime, please make this change as soon as possible.

As always, reach out to me if you have any questions or concerns.  Thanks
--
<')))><

David F. Chevrier

GIS / DevSecOps Lead, Northeast Fisheries Science Center